[BUGFIX] Link to cookie policy should be a page link, not action
When no cookies are detected, the link "Read more about our Cookie
Policy" is shown, which is linking to the action "editCookie".
This should be a link to a TYPO3 page, where the policy is explained and
where people can change their preferences, according the amount of
cookies we have. The policy page can be quite big, with a lot of text.
Also there is no way currently to define the policy itself.

The pid of the policy page should be configurable. On this policy page
the plugin "Cookie Control" can be placed.

This patch fixes this issue by adding editPid to constants and setup
which is used in the template for this link.

Fixes: #45479
Releases: 1.0
Change-Id: I8dbb60502ee8e62837b2989ad94298aee7fe2425
Reviewed-on: https://review.typo3.org/18298
Reviewed-by: Patrick Broens
Tested-by: Patrick Broens