[BUGFIX] AcceptCookie producing output when cookies have been set
The template AcceptCookie will always produce output, because the
condition is at the wrong spot. When a user has accepted/refused cookies
and the cookie_manager acceptance cookies have been set, the output
should be blank.

This patch fixes this by moving the condition and do some HTML cleaning

Fixes: #45481
Releases: 1.0
Change-Id: I905980b5c515edf5e13028fa38c4b974537b0510
Reviewed-on: https://review.typo3.org/18300
Reviewed-by: Patrick Broens
Tested-by: Patrick Broens